Corrección de warnings. Corrección en la generación de pasillos, las funciones de Generar pasillo vertical, y generar pasillo horizontal funcionan correctamente. Falta revisar que no colisionen los pasillos en esquinas, y calcular con las posiciones máximas un pasillo que no clisione. Corrección en la representación visual de los pasillos.

+++ LevelGenerator/LG_Utilities/Src/LG_Matrix4D.cpp
@@ -302,7 +302,7 @@ namespace LevelGenerator
 			}
 
 		///Calculamos la adjunta del lugar dado.
-		float fElevated = LG_Math::Pow(-1, ((col + 1) + (row + 1)));
+		float fElevated = LG_Math::Pow(-1.0f, (((float)col + 1.0f) + ((float)row + 1.0f)));
 		//TODO: corregir esto.
 		return fElevated * (fValues[0] * fValues[3] - fValues[1] * fValues[2]);
 	}
